LayoutTests/discovery/RUIHPL-7.9.3.1.1.h.html

<html>

<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>

<body>
<script>
var vid;
var fileName = '/home/gar/MediaContent/O-AVCMP4_BC15A-05.MP4';
var canplay = false;

function playable() {
	canplay = true;
	vid = document.getElementsByTagName("video")[0];
	vid.audioTracks.oncanplay = null;
	testPassed("Video Track is playable: " + fileName);
	finishJSTest();
}

function notPlayable()
{
	if (caplay)
		return;
	
	testFailed("Video Track is NOT playable: " + fileName);
	finishJSTest();
}

function insertCtrl() {
	var body = document.getElementsByTagName("body")[0];
	var v = document.createElement("video");
	v.setAttribute('controls', 'yes');
	v.setAttribute('width', '740px');
	v.addEventListener("canplay", playable, true);
	var src = document.createElement("source");
	src.setAttribute('src', fileName);
	v.appendChild(src);
	body.appendChild(v);
	setTimeout(notPlayable,4000);
}

insertCtrl();
description('Tests DLNA Player for AV Media Format. RUIHPL-7.9.3.1: AVC_MP4_MP_SD_AAC_MPS MP4 MP_L3 AAC MPS');
window.jsTestIsAsync = true;
</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>

</html>
